connect: move `ResolveContext` in own file
